+++ src/handler.ts
@@ -34,7 +34,8 @@ const handleFile = (response: File | Blob, set?: Context['set']) => {
 	) {
 		if (set) {
 			if (set.headers instanceof Headers)
-				if (hasHeaderShorthand) set.headers = set.headers.toJSON()
+				if (hasHeaderShorthand)
+					set.headers = (set.headers as unknown as Headers).toJSON()
 				else
 					for (const [key, value] of set.headers.entries())
 						if (key in set.headers) set.headers[key] = value
@@ -144,7 +145,7 @@ export const mapResponse = (
 			Array.isArray(set.headers['Set-Cookie'])
 		)
 			set.headers = parseSetCookies(
-				new Headers(set.headers),
+				new Headers(set.headers) as Headers,
 				set.headers['Set-Cookie']
 			) as any
 
@@ -182,7 +183,9 @@ export const mapResponse = (
 				const inherits = { ...set.headers }
 
 				if (hasHeaderShorthand)
-					set.headers = (response as Response).headers.toJSON()
+					set.headers = (
+						(response as Response).headers as Headers
+					).toJSON()
 				else
 					for (const [key, value] of (
 						response as Response
@@ -198,8 +201,9 @@ export const mapResponse = (
 				return errorToResponse(response as Error, set)
 
 			case 'Promise':
-				// @ts-ignore
-				return response.then((x) => mapResponse(x, set))
+				return (response as Promise<any>).then((x) =>
+					mapResponse(x, set)
+				) as any
 
 			case 'Function':
 				return mapResponse((response as Function)(), set)
@@ -218,7 +222,34 @@ export const mapResponse = (
 				return new Response(response?.toString(), set as SetResponse)
 
 			default:
+				if (response instanceof Response) {
+					const inherits = { ...set.headers }
+
+					if (hasHeaderShorthand)
+						set.headers = ((response as Response).headers as Headers).toJSON()
+					else
+						for (const [key, value] of (
+							response as Response
+						).headers.entries())
+							if (key in set.headers) set.headers[key] = value
+
+					for (const key in inherits)
+						(response as Response).headers.append(
+							key,
+							inherits[key]
+						)
+
+					return response as Response
+				}
+
+				if (response instanceof Promise)
+					return response.then((x) => mapResponse(x, set)) as any
+
+				if (response instanceof Error)
+					return errorToResponse(response as Error, set)
+
 				const r = JSON.stringify(response)
+
 				if (r.charCodeAt(0) === 123) {
 					if (!set.headers['Content-Type'])
 						set.headers['Content-Type'] = 'application/json'
@@ -294,7 +325,21 @@ export const mapResponse = (
 				return new Response(response?.toString(), set as SetResponse)
 
 			default:
+				if (response instanceof Response)
+					return new Response(response.body, {
+						headers: {
+							'Content-Type': 'application/json'
+						}
+					})
+
+				if (response instanceof Promise)
+					return response.then((x) => mapResponse(x, set)) as any
+
+				if (response instanceof Error)
+					return errorToResponse(response as Error, set)
+
 				const r = JSON.stringify(response)
+
 				if (r.charCodeAt(0) === 123)
 					return new Response(JSON.stringify(response), {
 						headers: {
@@ -350,7 +395,7 @@ export const mapEarlyResponse = (
 			Array.isArray(set.headers['Set-Cookie'])
 		)
 			set.headers = parseSetCookies(
-				new Headers(set.headers),
+				(new Headers(set.headers)) as Headers,
 				set.headers['Set-Cookie']
 			) as any
 
@@ -431,6 +476,32 @@ export const mapEarlyResponse = (
 				return new Response(response?.toString(), set as SetResponse)
 
 			default:
+				if (response instanceof Response) {
+					const inherits = { ...set.headers }
+
+					if (hasHeaderShorthand)
+						set.headers = ((response as Response).headers as Headers).toJSON()
+					else
+						for (const [key, value] of (
+							response as Response
+						).headers.entries())
+							if (key in set.headers) set.headers[key] = value
+
+					for (const key in inherits)
+						(response as Response).headers.append(
+							key,
+							inherits[key]
+						)
+
+					return response as Response
+				}
+
+				if (response instanceof Promise)
+					return response.then((x) => mapEarlyResponse(x, set)) as any
+
+				if (response instanceof Error)
+					return errorToResponse(response as Error, set)
+
 				const r = JSON.stringify(response)
 				if (r.charCodeAt(0) === 123) {
 					if (!set.headers['Content-Type'])
@@ -503,6 +574,19 @@ export const mapEarlyResponse = (
 				return new Response(response?.toString(), set as SetResponse)
 
 			default:
+				if (response instanceof Response)
+					return new Response(response.body, {
+						headers: {
+							'Content-Type': 'application/json'
+						}
+					})
+
+				if (response instanceof Promise)
+					return response.then((x) => mapEarlyResponse(x, set)) as any
+
+				if (response instanceof Error)
+					return errorToResponse(response as Error, set)
+
 				const r = JSON.stringify(response)
 				if (r.charCodeAt(0) === 123)
 					return new Response(JSON.stringify(response), {
@@ -571,13 +655,9 @@ export const mapCompactResponse = (response: unknown): Response => {
 
 		case 'Promise':
 			// @ts-ignore
-			return (response as any as Promise<unknown>).then((x) => {
-				const r = mapCompactResponse(x)
-
-				if (r !== undefined) return r
-
-				return new Response('')
-			})
+			return (response as any as Promise<unknown>).then(
+				mapCompactResponse
+			)
 
 		// ? Maybe response or Blob
 		case 'Function':
@@ -588,6 +668,19 @@ export const mapCompactResponse = (response: unknown): Response => {
 			return new Response((response as number | boolean).toString())
 
 		default:
+			if (response instanceof Response)
+				return new Response(response.body, {
+					headers: {
+						'Content-Type': 'application/json'
+					}
+				})
+
+			if (response instanceof Promise)
+				return response.then(mapCompactResponse) as any
+
+			if (response instanceof Error)
+				return errorToResponse(response as Error)
+
 			const r = JSON.stringify(response)
 			if (r.charCodeAt(0) === 123)
 				return new Response(JSON.stringify(response), {