+++ src/type-system.ts
@@ -44,7 +44,7 @@ try {
 		'date-time',
 		(value) => !Number.isNaN(new Date(value).getTime())
 	)
-} catch (error) {
+} catch {
 	// Not empty
 }
 
@@ -219,13 +219,16 @@ export const ElysiaType = {
 	) =>
 		t
 			.Transform(
-				t.Union([
-					t.String({
-						format: 'ObjectString',
-						default: ''
-					}),
-					t.Object(properties, options)
-				], options)
+				t.Union(
+					[
+						t.String({
+							format: 'ObjectString',
+							default: ''
+						}),
+						t.Object(properties, options)
+					],
+					options
+				)
 			)
 			.Decode((value) => {
 				if (typeof value === 'string')